pyproject.toml
[tool.poetry]
name = "heromatch"
version = "0.1.0"
description = ""
authors = ["michael <michael@mic-rigaud.fr>"]

[tool.poetry.dependencies]
python = "^3.10"
peewee = "^3.15.4"
invoke = "^2.0.0"
python-telegram-bot = {extras = ["job-queue"], version = "^20.4"}
Pillow = "^10.1.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
